+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # WaterDrop changelog
 
 ## 2.7.3 (Unreleased)
+- [Enhancement] Introduce `reload_on_transaction_fatal_error` to reload the librdkafka after transactional failures
+- [Enhancement] Flush on fatal transactional errors.
 - [Enhancement] Add topic scope to `report_metric` (YadhuPrakash)
 - [Enhancement] Cache middleware reference saving 1 object allocation on each message dispatch.
 - [Enhancement] Provide `#idempotent?` similar to `#transactional?`.